Rewrite using a more top down approach
Trying to be all super TDD and only do bare minimum for each test
but its taking 10 times longer because I keep hitting dead ends
and going down a path that I basically think is shitty. Eventually
hit a spot where I couldn't really see a way to go on, so rewriting
using top-down, less functional methods (methods on the module that
I pass self to and so forth), more objects with responsibilities.