Initialize is no longer implicitly recorded.
That was adding such ugly code, and you wound up often having to define it
anyway, because otherwise you would have arity issues after recording

It would be nice if you wanted it recorded to be able to call super, but
it is simply not possible at the moment due to implementation (I doubt
there is an alternative implementation that will satisfy the
requirements, and allow this behaviour)

Instead, maybe try to add an ability to declare that methods be
recorded, independently of specifying default block. Maybe the
default block is simply the method being called.