Get params from the method that the user defined instead of from the method surrogate defined.
This is really gross, but I just need to tie up some of these loose ends
before I try to refactor it. After this is all working, go through and
figure out the right way to do all of this (this information already
exists somewhere, e.g. to raise the argument error we must know this.